requirements.txt
# Core ML
torch==2.3.1
torchvision==0.18.1

# Data + utils
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
opencv-python==4.10.0.84
scikit-learn==1.5.1
matplotlib==3.9.0
seaborn==0.13.2
tqdm==4.66.4

# Serving
streamlit==1.36.0

# For downloading datasets
kaggle==1.6.14
